backend/requirement.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
decouple==0.0.7
Django==4.2.7
django-cors-headers==4.3.0
django-filter==23.3
djangorestframework==3.14.0
idna==3.4
Markdown==3.5.1
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
urllib3==2.0.7
